requirements.txt
requests
python-Levenshtein